fix: telemetry for closed ports after debug finish
