+++ packages/vscode-extension/src/debug/teamsfxDebugProvider.ts
@@ -187,7 +187,7 @@ export class TeamsfxDebugProvider implements vscode.DebugConfigurationProvider {
       debugConfiguration.teamsfxResolved = true;
     } catch (error: any) {
       void showError(error);
-      await terminateAllRunningTeamsfxTasks();
+      terminateAllRunningTeamsfxTasks();
       await vscode.debug.stopDebugging();
       // not for undefined
       if (telemetryIsRemote === false) {
