+++ packages/vscode-extension/src/debug/teamsfxTaskHandler.ts
@@ -45,7 +45,7 @@ import {
 } from "./common/localDebugSession";
 import { allRunningDebugSessions } from "./officeTaskHandler";
 import { deleteAad } from "./deleteAadHelper";
-import detectPort from "detect-port";
+
 class NpmInstallTaskInfo {
   private startTime: number;
 
@@ -231,7 +231,7 @@ async function onDidStartTaskProcessHandler(event: vscode.TaskProcessStartEvent)
       const session = getLocalDebugSession();
       if (session.id !== DebugNoSessionId) {
         if (session.failedServices.length > 0) {
-          await terminateAllRunningTeamsfxTasks();
+          terminateAllRunningTeamsfxTasks();
           // TODO: send test tool log file
           await sendDebugAllEvent(
             new UserError({
@@ -282,7 +282,7 @@ async function onDidEndTaskProcessHandler(event: vscode.TaskProcessEndEvent): Pr
       // This is the final task for test tool.
       // If this task exits (even exitCode is 0) before being successfully started, the debug fails.
       if (currentSession.id !== DebugNoSessionId) {
-        await terminateAllRunningTeamsfxTasks();
+        terminateAllRunningTeamsfxTasks();
         await sendDebugAllEvent(
           new UserError({
             source: ExtensionSource,
@@ -319,7 +319,7 @@ async function onDidEndTaskProcessHandler(event: vscode.TaskProcessEndEvent): Pr
     event.exitCode !== 0 &&
     event.exitCode !== -1
   ) {
-    await terminateAllRunningTeamsfxTasks();
+    terminateAllRunningTeamsfxTasks();
   } else if (isNpmInstallTask(task)) {
     try {
       const taskInfo = activeNpmInstallTasks.get(task.name);
@@ -439,7 +439,7 @@ async function onDidEndTaskProcessHandler(event: vscode.TaskProcessEndEvent): Pr
             )
           );
         }
-        await terminateAllRunningTeamsfxTasks();
+        terminateAllRunningTeamsfxTasks();
       }
     } catch {
       // ignore any error
@@ -499,7 +499,7 @@ async function onDidStartDebugSessionHandler(event: vscode.DebugSession): Promis
         // Handle cases that some services failed immediately after start.
         const currentSession = getLocalDebugSession();
         if (currentSession.id !== DebugNoSessionId && currentSession.failedServices.length > 0) {
-          await terminateAllRunningTeamsfxTasks();
+          terminateAllRunningTeamsfxTasks();
           await vscode.debug.stopDebugging();
           await sendDebugAllEvent(
             new UserError({
@@ -524,7 +524,7 @@ async function onDidStartDebugSessionHandler(event: vscode.DebugSession): Promis
   }
 }
 
-export async function terminateAllRunningTeamsfxTasks(): Promise<void> {
+export function terminateAllRunningTeamsfxTasks(): void {
   for (const task of allRunningTeamsfxTasks) {
     try {
       if (task[1] > 0) {
@@ -534,20 +534,12 @@ export async function terminateAllRunningTeamsfxTasks(): Promise<void> {
       // ignore and keep killing others
     }
   }
-  const closedPorts: number[] = [];
-  for (const port of globalVariables.LocalDebugPorts.checkPorts) {
-    const port2 = await detectPort(port);
-    if (port2 === port) {
-      closedPorts.push(port);
-    }
-  }
-  globalVariables.LocalDebugPorts.closedPorts = closedPorts;
   allRunningTeamsfxTasks.clear();
   BaseTunnelTaskTerminal.stopAll();
   void deleteAad();
 }
 
-async function onDidTerminateDebugSessionHandler(event: vscode.DebugSession): Promise<void> {
+function onDidTerminateDebugSessionHandler(event: vscode.DebugSession): void {
   if (allRunningDebugSessions.has(event.id)) {
     // a valid debug session
     // send stop-debug event telemetry
@@ -555,7 +547,7 @@ async function onDidTerminateDebugSessionHandler(event: vscode.DebugSession): Pr
       [TelemetryProperty.DebugSessionId]: event.id,
     });
 
-    await terminateAllRunningTeamsfxTasks();
+    terminateAllRunningTeamsfxTasks();
 
     allRunningDebugSessions.delete(event.id);
     if (allRunningDebugSessions.size == 0) {
